+++ My_Debtors_Backend/my_debtors_api/models.py
@@ -23,9 +23,11 @@ class Student(models.Model):
         return self.first_name
 
 class Debtor(models.Model):
-    user_id = models.ForeignKey(User, on_delete=models.CASCADE)
+    user_id = models.ForeignKey(User, on_delete=models.CASCADE, null=True)
     school_id = models.ForeignKey(School, on_delete=models.CASCADE)
     student_id = models.ForeignKey(Student, on_delete=models.CASCADE)
+    first_name = models.CharField(max_length=20)
+    last_name = models.CharField(max_length=20)
     phone_number = models.IntegerField()
     resident_address = models.CharField(max_length=100)
     amount_owed = models.IntegerField()
