+++ My_Debtors_Backend/my_debtors_api/serializers.py
@@ -1,5 +1,5 @@
 from rest_framework.serializers import ModelSerializer
-from my_debtors_api.models import Debtor
+from .models import Debtor
 
 class DebtorSerializer(ModelSerializer):
     class Meta:
