+++ My_Debtors_Backend/my_debtors_core/urls.py
@@ -4,5 +4,6 @@ from django.urls import path, include
 urlpatterns = [
     path("admin/", admin.site.urls),
     path('auth/user/', include('authentication.urls')),
-    path('api/', include('my_debtors_api.debtors_api.urls'))
+    path('api-auth/', include('rest_framework.urls')),
+    path('api/', include('my_debtors_api.urls'))
 ]
